Improve training stability and performance in segmentation model
- Added handling for non-finite loss values to skip problematic batches, preventing weight poisoning.
- Changed loss calculation to ensure it is computed in float32 to avoid overflow issues on T4 GPUs.
- Updated progress bar to display loss every 20 batches instead of every iteration for better performance.
- Optimized gradient clipping and scaling operations for improved training efficiency.

+++ segmentation_model/eval.py
@@ -77,7 +77,7 @@ def main():
         for i, batch in enumerate(dl):
             x = batch["image"].to(device)
             y = batch["label"].to(device)
-            with torch.amp.autocast(device, enabled=device == "cuda"):
+            with torch.amp.autocast(device, dtype=torch.bfloat16, enabled=device == "cuda"):
                 out = sliding_window_inference(x, (CROP, CROP), 2, model)
             p = [post_pred(o) for o in out]
             t = [post_lbl(o) for o in y]
